First set of changes to enable core-os cloud config
Added a couple of new parameters:

- volid: Needs to be set to "config-2" for coreos, defaults to
  "cidata"
- user_data_path: Needs to be set to "openstack/latest/user_data"
  for coreos, defaults to "user-data"

The following parameter is still being developed, but should be
safe to use with for instance the terraform file parameter to
read a cloud-init file into a string.

- user_data: This will be used to consume a cloud-init or coreos
  specific configuration file generated elsewhere by terraform.

The user_data parameter has been very lightly tested with the
terraform cloud-init template, and seems to work to the extent
that rendered templates are decoded from base64 if base64 decoding
is set.